DP-7022 minor theme helpers improvements
